Add files via upload
Budget Planning and Tracking Program
This program, developed in Excel, helps users efficiently plan their budget, track financial transactions, and analyze financial outcomes through an interactive dashboard.

Functionality Overview
1. "Budget Planning" Sheet
Planned Income, Expenses, and Savings:
In the green section, planned income is entered.
In the red section, planned expenses are specified.
In the blue section, planned savings are displayed.
Planning Dynamics:
In the "Planning Dynamics" column, the sums of income, expenses, and savings must equal zero. If there is a discrepancy, the program will indicate the amount that is yet to be planned, allowing for adjustments.
2. "Budget Tracking" Sheet
Financial Transactions:
The first block shows the current day.
The second block displays the date of the last recorded transaction.
The third block calculates how many days have passed since the last recorded transaction.
Transaction Table:
This table logs transactions for income, expenses, and savings, which automatically generate data for the future dashboard.
3. "Dashboard" Sheet
Period Analysis:
The period is selected using a dropdown filter.
The percentage of the elapsed time for the selected period is displayed.
Data Visualization:
The percentage of plan completion for income, expenses, and savings is shown.
A graph displays the Top-6 positions by amount in income, expenses, and savings.
A bar chart shows the total amount of expenses for each category (income, expenses, and savings).